requirements.txt
Flask==3.0.3
Flask-SocketIO==5.3.0
python-dotenv==1.1.1
requests==2.32.5